Move some code out into versioned modules
